AI_devs4/se02e05/pyproject.toml
[project]
name = "drone-mission"
version = "0.1.0"
description = "Two-agent drone mission system"
requires-python = ">=3.13"
dependencies = [
    "dspy[mcp]>=3.1.3",
    "python-dotenv>=1.2.2",
    "requests>=2.32.5",
    "rich>=14.3.3",
    "ruff>=0.15.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
drone = "src.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src"]
